requirements.txt
# langgraph==1.0.5
# langgraph-prebuilt==1.0.5
# python-dotenv
# langsmith
# langchain-openai
# langchain-community
# langchain-tavily
# langchain-experimental
# tavily-python 
# pandas 
requests
azure-identity
IPython
#azure-ai-agentserver-agentframework
azure-ai-agentserver-agentframework==1.0.0b7
agent-framework
#azure-ai-agentserver-langgraph==1.0.0b7
azure-monitor-opentelemetry==1.8.1
python-dotenv
openai
azure-ai-projects>=2.0.0a20250915020
opentelemetry-instrumentation-openai-v2